always save a checkpoint at start of training - no-op if already exists

+++ utils/checkpointing.py
@@ -68,32 +68,35 @@ def save_checkpoint(fabric, training_config, model, optimizer, lr_scheduler, ste
         return
 
     run_dir = os.path.join(RUNS_DIR, training_config.run_name)
+
     root_checkpoint_dir = os.path.join(run_dir, CHECKPOINT_DIR)
     os.makedirs(root_checkpoint_dir, exist_ok=True)
 
     curr_checkpoint_dir = os.path.join(root_checkpoint_dir, f"step_{step}")
-
     os.makedirs(curr_checkpoint_dir, exist_ok=True)
 
     model_state_path = os.path.join(curr_checkpoint_dir, "model.pt")
+    if not os.path.exists(model_state_path):
+        model_state = {
+            "model": model,
+        }
+        fabric.save(model_state_path, model_state)
+
     optimizer_state_path = os.path.join(curr_checkpoint_dir, "optimizer.pt")
+    if not os.path.exists(optimizer_state_path):
+        optimizer_state = {
+            "optimizer": optimizer,
+            "lr_scheduler": lr_scheduler,
+        }
+        fabric.save(optimizer_state_path, optimizer_state)
+
     training_state_path = os.path.join(curr_checkpoint_dir, "training.pt")
-
-    model_state = {
-        "model": model,
-    }
-    optimizer_state = {
-        "optimizer": optimizer,
-        "lr_scheduler": lr_scheduler,
-    }
-    training_state = {
-        "step": step,
-        "rng_state": _collect_rng_states(),
-    }
-
-    fabric.save(model_state_path, model_state)
-    fabric.save(optimizer_state_path, optimizer_state)
-    fabric.save(training_state_path, training_state)
+    if not os.path.exists(training_state_path):
+        training_state = {
+            "step": step,
+            "rng_state": _collect_rng_states(),
+        }
+        fabric.save(training_state_path, training_state)
 
     # create symlink to latest checkpoint directory
     latest_symlink_path = os.path.join(run_dir, CHECKPOINT_DIR, "latest")
@@ -103,6 +106,7 @@ def save_checkpoint(fabric, training_config, model, optimizer, lr_scheduler, ste
     os.symlink(f"step_{step}", latest_symlink_path, target_is_directory=True)
 
     # Pushing to HuggingFace Hub
+    # NOTE: if the file already exists, HF will not upload it again (by default)
     if training_config.checkpointing.hf_repo_id is not None:
         if step == 0:
             # upload the config to the HuggingFace Hub
@@ -146,9 +150,10 @@ def save_config(fabric, training_config, model_config, evaluation_config):
         return
 
     config_path = os.path.join(RUNS_DIR, training_config.run_name, "config.yaml")
-    with open(config_path, "w") as f:
-        yaml.dump(training_config, f)
-        yaml.dump(model_config, f)
-        yaml.dump(evaluation_config, f)
+    if not os.path.exists(config_path):
+        with open(config_path, "w") as f:
+            yaml.dump(training_config, f)
+            yaml.dump(model_config, f)
+            yaml.dump(evaluation_config, f)
 
     fabric.barrier()
